Add pictures of 2020 day 8 control flow graph.
It's getting a little ridiculous, but since it's possible to do a
control flow graph representation of the day 8 program (showing both
the as-is and potentially repaired instructions), let's do that.

This also well illustrates the fact that there's a potentially more
elegant solution available: treating the graph as weighted, with edge
weights of 0 for the original instruction and 1 for its potential
replacement, find the shortest path to the halting node. This will
immediately indicate which instruction needs to be repaired.

Actually implementing that solution will come later, if ever.