Refactor the glue code one more time.
This commit moves the glue out of the `util` package (which was a bad
fit anyway), and adds years as part of the puzzle identifier, so that
everything can be bundled into a single binary. It also adds a
registration system for dot plotters, in preparation for adding
quality-of-life features like direct-to-image output and regenerating
the docs plots.

The I/O code in `util` is also harmonized a little, including adding a
`ScanAll` function to run a scanner to completion on an io.Reader.