Solve AoC 2018 day 23 (badly).
This solution is terribly slow. It uses an octree branch-and-bound
thing, where for each cube the bound is based on the total number of
bots that could possibly be in reach anywhere in the cube, as well as
the minimum distance between the cube and origin: exploration is not
continued if those bounds show there is no chance of getting a better
result than the current best known. Yet it still takes ages (>6s).