Improve 2018 day 23 by 10x or more.
Sigh, should've seen that immediately. The key insight in this commit
is that, if one of the octree cubes is entirely inside the range of a
bot, we need not include it in the list of filtered bots. We can just
pass along a base offset of the number of overlapping bots. This way
all the exploration near the boundary has a much, much shorter list of
bots to distance-test. This single change brings the runtime for my
puzzle input down from over six seconds to under half.