Refactor solver glue code.
Previously, solvers in general returned a `[]int`, requiring to use
the special `glue.GenericSolver` type (with the input parsing repeated
in the solver) when textual output was needed. This commit inverts
that: solvers now return a `[]string`, but the glue code exports a
variadic `glue.Ints` function to make returning some amount of
integers convenient.

It also adds a `glue.RegexpSolver` convenience type for solutions that
need to apply a specific regular expression to each line of the input
in order to parse it. Only year 2017 solutions for now are migrated to
use it.