Remove an unnecessary layer of indirection from 2020 day 23.
Previously we allocated each cup independently, and used a slice of
pointers to the nodes to look them up by label. But there's no reason
why we couldn't simply use a slice of the node structures, and take
its elements' addresses when we need pointers. This simple change cuts
the runtime almost by half.