Optimize 2020 day 23 one more step.
After the previous change, each list node's index is always the same
as its label. So we can forget about the rest of the node structure,
and simply store the `next` links. Making them array indices means the
same value serves as both the label and the pointer. This cuts the
runtime still further, to about a third of the original.