Implementation improvements on 2020 day 22.
These cut the runtime to about half of what it used to be. The big win
is from changing the card array to be a `[]byte` rather than an
`[]int`, which also allows writing it (almost) directly to the FNV
hash. It's still the hashing and the hash map checking that takes up
most of the runtime though.