Downsample stats-based data visualizations.
Keeping statistics at the 5-minute granularity for all 25 days over
the whole month was making the generated charts quite bulky. Generally
the numbers will not change as rapidly once the puzzle is old. This
commit keeps the 5-minute data for the first day, downsamples to
hourly data for the first week, and to daily data for the rest of the
contest.