Bulesque improvements to past 2021 days.
Turns out there's a pair of builtins, `g_` and `l_`, which normally do
gcd and lcm on integers, but on lists pop off one element from either
end, and leave on stack the remainder of the list plus that one
element. This streamlines a lot of things. There are probably other
days this could apply to, too.