Miscellaneous visualization improvements.
Stop trying to sync day 26 even if the script is executed during one
of the Dec 26-31 days. Add `byday` plots to illustrate the usual
puzzle difficulty (or is it interest?) curve. Filter ratios to avoid
zeros, which don't work on the log-scale charts. Add a heatmap style
version of the Go benchmark times.