Add a bitmap utility.
A lot of the puzzles need a map of visited coordinates. Sometimes the
grid is of fixed size, but occasionally it's not. This commit adds a
basic two-dimensional bitmap made out of 8x8 (`uint64`) blocks of
bits, with a small fixed-size (64x64) "core" augmented by a map. It
gives a modest speed boost to the 2022 day 9 puzzle.