Performance improvements for 2022 day 16.
Still no real breakthrough here, but these changes at least get the
runtime to 0.55 seconds, under the psychologically important
one-second boundary.

On the implementation front, this changes the "log" (used to store the
lists of achievable time-pressure scores, keyed by current locations
and so-far opened vents) from a map into just a large array. It's
maybe a little extravagant for part 2 (where the size is 15*15*65536
elements), but does seem to benchmark better than the map anyway.

In terms of the algorithm, the only major win is adding a new pruning
rule. This is based on comparing an upper bound of the maximum
achievable released pressure (opening all the remaining vents no more
than what the minimum travel time between any two vents allows) to the
current best result, and not expanding a path if it can't beat that.