Solve AoC 2022 day 19 (also quite badly).
This commit also includes some incidental tweaks to day 16 to bring it
under the unwritten 0.5s bound, meaning it gets to be included in the
Go benchmark visualizations again. Day 19 also just barely makes it.

For both, there's almost certainly some clever trick to leverage. It's
just not occurring to me.