Add a silly queue type.
After getting this idea of using the `len` field of a slice as a
general-purpose data field instead of the normal slice semantics,
couldn't help myself but to implement it. It's probably objectively a
bad idea, but it's just so silly.

There's some borderline reasonable arguments for not using a slice
with the `s = s[1:]` pop operation as a queue, since it involves
continuously extending a slice, and therefore does more copying (and
can take time free up the storage) than strictly necessary. This
implementation uses a typical ring buffer instead, which should be
zero-copy as long as the queue size does not grow indefinitely. Any
performance difference is probably negligible or even negative, but it
is what it is.