Redo the graph utility library.
The previous thing was a very once-size-fits-all solution where
everything was a dense adjacency-matrix representation. Most AoC
puzzles tend to have sparse graphs and only require iteration in the
"forward" direction, so unidirectional edge lists are perfectly
adequate. Adding a thing that supports different graph types with
(mostly) the same API, so it'll be hopefully easier to compare and
contrast the efficiency.