Simplify writing generic graph functions.
Instead of a generics nonsense to deal with the different iterator
types, use a callback-based iteration instead.

Rumor has it that it can be just as fast as iterators (due to
mid-stack inlining), though some microbenchmarks suggest that's not
really happening here. Still, the generics were even worse. Also tried
making a single unified iterator type, but it wasn't as good as the
dedicated ones.